+++ packages/lumx-react/src/components/checkbox/Checkbox.tsx
@@ -15,7 +15,7 @@ import uniqueId from 'lodash/uniqueId';
  */
 interface CheckboxProps extends GenericProps {
     /** Whether it is checked or not. */
-    checked?: boolean;
+    isChecked?: boolean;
     /** Is checkbox disabled */
     isDisabled?: boolean;
     /** Helper */
@@ -33,7 +33,7 @@ interface CheckboxProps extends GenericProps {
     /** Native input value. */
     value?: string;
     /** Handle onChange event. */
-    onChange?(checked: boolean, value?: string, name?: string, event?: SyntheticEvent): void;
+    onChange?(isChecked: boolean, value?: string, name?: string, event?: SyntheticEvent): void;
 }
 
 /**
@@ -62,9 +62,10 @@ const Checkbox: React.FC<CheckboxProps> = ({
     checked,
     className,
     disabled,
-    isDisabled = disabled,
     helper,
     id,
+    isChecked = checked,
+    isDisabled = disabled,
     label,
     name,
     onChange,
@@ -76,7 +77,7 @@ const Checkbox: React.FC<CheckboxProps> = ({
     const inputId = id || uniqueId(`${CLASSNAME.toLowerCase()}-`);
     const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
         if (onChange) {
-            onChange(!checked, value, name, event);
+            onChange(!isChecked, value, name, event);
         }
     };
 
@@ -86,9 +87,9 @@ const Checkbox: React.FC<CheckboxProps> = ({
             className={classNames(
                 className,
                 handleBasicClasses({
-                    isChecked: checked,
+                    isChecked,
                     isDisabled,
-                    isUnchecked: !checked,
+                    isUnchecked: !isChecked,
                     prefix: CLASSNAME,
                     theme,
                 }),
@@ -103,7 +104,7 @@ const Checkbox: React.FC<CheckboxProps> = ({
                     tabIndex={isDisabled ? -1 : 0}
                     name={name}
                     value={value}
-                    checked={checked}
+                    checked={isChecked}
                     onChange={handleChange}
                 />