+++ packages/lumx-react/src/components/switch/Switch.tsx
@@ -20,7 +20,7 @@ enum SwitchPosition {
  */
 interface SwitchProps extends GenericProps {
     /** Whether it is checked or not. */
-    checked?: boolean;
+    isChecked?: boolean;
 
     /** Switch disabled state. */
     isDisabled?: boolean;
@@ -50,7 +50,7 @@ interface SwitchProps extends GenericProps {
     value?: string;
 
     /** Handle onChange event. */
-    onChange?(checked: boolean, value?: string, name?: string, event?: SyntheticEvent): void;
+    onChange?(isChecked: boolean, value?: string, name?: string, event?: SyntheticEvent): void;
 }
 
 /**
@@ -82,6 +82,7 @@ const Switch: React.FC<SwitchProps> = ({
     className,
     disabled,
     helper,
+    isChecked = checked,
     isDisabled = disabled,
     name,
     onChange,
@@ -94,7 +95,7 @@ const Switch: React.FC<SwitchProps> = ({
     const switchId: string = uuid();
     const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
         if (onChange) {
-            onChange(!checked, value, name, event);
+            onChange(!isChecked, value, name, event);
         }
     };
 
@@ -105,11 +106,11 @@ const Switch: React.FC<SwitchProps> = ({
                 className,
                 handleBasicClasses({
                     prefix: CLASSNAME,
-                    checked: Boolean(checked),
+                    isChecked,
                     isDisabled,
                     position,
                     theme,
-                    unchecked: !Boolean(checked),
+                    isUnchecked: !isChecked,
                 }),
                 { [`${CSS_PREFIX}-custom-colors`]: useCustomColors },
             )}
@@ -123,7 +124,7 @@ const Switch: React.FC<SwitchProps> = ({
                     name={name}
                     value={value}
                     disabled={isDisabled}
-                    checked={checked}
+                    checked={isChecked}
                     onChange={handleChange}
                 />