+++ packages/site-demo/content/product/components/radio-button/react/default.tsx
@@ -8,7 +8,7 @@ export const App = ({ theme }: any) => {
     return (
         <RadioGroup>
             <RadioButton
-                checked={value === 'lorem'}
+                isChecked={value === 'lorem'}
                 label="Radio button"
                 name="test1"
                 theme={theme}
@@ -17,7 +17,7 @@ export const App = ({ theme }: any) => {
             />
 
             <RadioButton
-                checked={value === 'ipsum'}
+                isChecked={value === 'ipsum'}
                 helper="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed posuere faucibus efficitur."
                 label="Radio button with help"
                 name="test1"
@@ -27,7 +27,7 @@ export const App = ({ theme }: any) => {
             />
 
             <RadioButton
-                checked={value === 'dolor'}
+                isChecked={value === 'dolor'}
                 helper="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed posuere faucibus efficitur."
                 label="Disabled radio button with help"
                 name="test1"